+++ .github/workflows/build.sh
@@ -9,8 +9,8 @@ cargo install comrak
 mkdir -p dist
 cp -r target/llvm-cov/html dist/coverage
 sed 's|CHANGELOG.md|CHANGELOG.html|g' README.md >README.md.tmp
-comrak --gfm -o dist/index.html --width 10 README.md.tmp
-comrak --gfm -o dist/CHANGELOG.html --width 10 CHANGELOG.md
+comrak --gfm -o dist/index.html --width 80 README.md.tmp
+comrak --gfm -o dist/CHANGELOG.html --width 80 CHANGELOG.md
 
 # cleanup
 rm README.md.tmp